+++ packages/gator-permissions-controller/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add `submitRevocation` action to submit permission revocations through the gator permissions provider snap
 - Add `addPendingRevocation` action to queue revocations until transaction confirmation
+- Expose list of pending revocations in state
 
 ### Changed